Explicitly use conformer_id when getting generated conformers
`mol_from_ccd_cif` adds the ideal conformer from the ccd cif before we create a generated conformer. Because `params.clearConf` is True by default this is cleared, which means the default conformer we we currently get is correctly the randomly generated one not the ideal conformer. However, we now use the returned `conformer_id` explicitly to be safe in case of future changes.

PiperOrigin-RevId: 701962101